为删除话题增加安全认证

+++ controllers/topic.js
@@ -75,18 +75,40 @@ exports.edit = (req,res,next) => {
 	res.send('post edit')
 }
 
+// 处理删除请求
 exports.delete = (req,res,next) => {
-	// res.send('post delete')
 	const {topicId} = req.params
 	// console.log(topicId)
-	topic.findByIdAndRemove(topicId,(err,results) => {
+	// 根据当前的话题ID查询当前登录的用户ID
+	topic.findById(topicId,(err,result) => {
 		if(err){
-			return next(err)
+			return next(err)    // 服务端错误
 		}
-
-		res.status(200).json({
-			code: 0,
-			message: 'success'
+		// 如果当前话题不存在
+		if(!result){
+			return res.status(200).json({
+				code: 1,
+				message: '对不起，该资源已不存在！'
+			})
+		}
+		// 如果服务端没有错误并查到了当前话题 信息
+		// 判断当前话题的用户id是不是当前登录用户的id
+		if(result.userId !== req.session.user.id){
+			return res.status(200).json({
+				code: 2,
+				message: '对不起，您没有权限删除该话题'
+			})
+		}
+		// 删除话题
+		topic.findByIdAndRemove(topicId,(err,results) => {
+			if(err){
+				return next(err)
+			}
+			return res.status(200).json({
+				code: 0,
+				message: 'success'
+			})
 		})
 	})
+	
 }
